refactor: Extract InputField component for reusability
