+++ CVCraft/src/components/InfoForm/shared/InputField.jsx
@@ -1,0 +1,17 @@
+import PropTypes from "prop-types";
+
+function InputField({ fieldId, value, onChange }) {
+    return <input type={"text"}
+                  id={fieldId}
+                  name={fieldId}
+                  value={value}
+                  onChange={onChange} />;
+}
+
+InputField.propTypes = {
+    fieldId: PropTypes.string.isRequired,
+    value: PropTypes.string.isRequired,
+    onChange: PropTypes.func.isRequired
+};
+
+export default InputField;
