+++ CVCraft/src/styles/InfoForm/PersonalData/personal_data_style.css
@@ -2,6 +2,8 @@
     margin: 0;
     padding: 0;
     box-sizing: border-box;
+
+    font-size: 15px;
 }
 
 .personal-info > ul {
@@ -14,7 +16,7 @@ li {
 }
 
 li > label {
-    display: block;
+    display: inline-block;
     margin-bottom: 5px;
 }
